updated BookInfoScreenViewModel.kt
